app/requirements.txt
Pillow~=7.2.0
Flask~=1.1.2
pytesseract~=0.3.4
azure-storage-blob~=12.3.2